Decompose logically disparate groups into their own classes/files
